logpersist: Inserting and unplugging U disk causes black screen [1/1]
PD#SWPL-185854

Problem:
kill invalid Pid

Solution:
Support exfat format U disk

Verify:
qurra

+++ frameworks/services/logging/UserLogConfig.cpp
@@ -111,7 +111,10 @@ pid_t UserLogConfig::startPersistLogcat() {
     return pid;
 }
 bool UserLogConfig::stopLogging() {
-   return kill(mLogcatConfig.logcat_pid, 9) >= 0;
+    if (mLogcatConfig.logcat_pid != -1)
+        return kill(mLogcatConfig.logcat_pid, 9) >= 0;
+    else
+        return true;
 }
